Add content.css and content.min.css files
These new files include necessary styling for the content and panels of the site, as well as the form submissions. This affects various elements such as buttons, content selectors and submit buttons. In addition, this change implements specific styles for error and success states in forms which enhances the user interface.