Update website's layout and font
Modified the website's structure to make it more user-friendly, with better positioning of key information, and changed the website's font from Lato to Poppins for a fresher look. The changes also include adding new sections for HR Contact Information and modifying the presentation of FAQ along with an added functionality for printing FAQs.