Update banner font sizes in responsive.css
In the responsive.css and responsive.min.css files, the font sizes for various elements within the .banner class have been decreased. This includes the anchor tags and the h2 element. In addition, a minimum width has been set for the .col class within .banner to ensure readability on smaller screens.