Wrap kiosk.js and kiosk.min.js in an IIFE
Wrapped the entire kiosk.js and kiosk.min.js code in an Immediately Invoked Function Expression (IIFE). This change ensures that all the variables and functions defined in kiosk.js and kiosk.min.js are not globally scoped, improving code encapsulation and reducing the potential interference with other scripts on the page.