Update opacity and pointer-events for .darken class
The opacity for the .darken class in the CSS was increased from 0.1 to 0.5 to allow for more visibility. The "display: none" rule has been commented out and a "pointer-events: none" rule has been added to disable any mouse interactions. This change has been reflected in both the content.css and content.min.css files.